+++ packages/parte-icons/src/Icons/ArrowRightIcon.tsx
@@ -1,0 +1,30 @@
+import { convertIcon, IconProps } from "../common/Icon";
+function Icon({ size, ...rest }: Pick<IconProps, "style" | "color" | "size">) {
+  return (
+    <svg
+      width={size}
+      height={size}
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...rest}
+    >
+      <path
+        d="M14 8H2"
+        stroke="currentcolor"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M10 4L14 8L10 12"
+        stroke="currentcolor"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
+const ArrowRightIcon = convertIcon(Icon, "arrow-right");
+export default ArrowRightIcon;