+++ apps/web/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @parte-ds/ui@1.14.1
   - @parte-ds/icons@1.7.0
 
 ## 0.1.33